src/requirements.txt
django==4.2
gunicorn
requests
python-decouple
dj-database-url
psycopg2-binary
djangorestframework
markdown
django-filter
PyYAML
uritemplate